refactor(cli): consolidate the package manager resolver and project file iteration
