+++ .changeset/cli-shared-utils.md
@@ -1,0 +1,5 @@
+---
+"assistant-ui": patch
+---
+
+refactor: consolidate the package manager resolver and project file iteration
